PR Feedback - switch to MedicationAdministration resource
