+++ data/Templates/eCR/Resource/Composition.liquid
@@ -257,10 +257,10 @@
                                     {% endfor %}
                                 {% elsif sectionEntry.substanceAdministration != nil and component.section.templateId.root == '2.16.840.1.113883.10.20.22.2.80' %}
                                     {% comment %} Pregnancy substanceAdministrations {% endcomment %}
-                                    {% assign observationId = sectionEntry.substanceAdministration | to_json_string | generate_uuid -%}
-                                    {% assign fullObservationId = observationId | prepend: 'Observation/' %}
+                                    {% assign substanceAdministrationId = sectionEntry.substanceAdministration | to_json_string | generate_uuid -%}
+                                    {% assign fullSubstanceAdministrationId = substanceAdministrationId | prepend: 'MedicationAdministration/' %}
                                     {
-                                        "reference": "{{ fullObservationId }}",
+                                        "reference": "{{ fullSubstanceAdministrationId }}",
                                     },
                                 {% endif -%}
                             {% endfor -%}
