--- data/Templates/eCR/Resource/ObservationPregnancySubstanceAdministration.liquid
@@ -1,38 +1,0 @@
-{
-    "fullUrl":"urn:uuid:{{ ID }}",
-    "resource":{
-        "resourceType": "Observation",
-        "id":"{{ ID }}",
-         "meta":
-        {
-            "profile":
-            [
-                "http://hl7.org/fhir/StructureDefinition/Observation"
-            ],
-        },
-        "identifier":
-        [
-            {% assign ids = observationEntry.id | to_array -%}
-            {% for id in ids -%}
-                {% include 'DataType/Identifier', Identifier: id -%}
-            {% endfor -%}
-        ],
-        "status": "{{ observationEntry.statusCode.code | get_property: 'ValueSet/ObservationStatus' }}",
-        {% include 'Utils/EffectiveTime', effectiveTime: observationEntry.effectiveTime %}
-        "code":
-        {
-            "coding": [
-                {
-                    "system" : "http://loinc.org",
-                    "code" : "73813-8",
-                    "display": "Characteristics of labor and delivery [US Standard Certificate of Live Birth]"
-                }
-            ],
-            "text" : "Characteristics of labor and delivery"
-        },
-        {{ observationEntry.consumable.manufacturedProduct.manufacturedMaterial.code | print_object }}
-        {% if observationEntry.consumable.manufacturedProduct.manufacturedMaterial.code %}
-            {% include 'Utils/ValueHelper', value: observationEntry.consumable.manufacturedProduct.manufacturedMaterial.code -%}
-        {% endif %}
-    },
-},
